Added clearing cache on logout, added reseting redux on login

+++ src/Frontend/src/utils/localStorageManager.ts
@@ -1,4 +1,9 @@
+let blockCache = false;//when true blocking any changes in cache
+
 export function saveToCache<T>(groupName: string, id: string, value: T): void {
+	if(blockCache) {
+		return;
+	}
 	const data = JSON.parse(localStorage[groupName] || '{}');
 	data[id] = value;
 	localStorage[groupName] = JSON.stringify(data);
@@ -10,11 +15,25 @@ export function loadFromCache<T>(groupName: string, id: string): T | undefined {
 }
 
 export function removeFromCache(groupName: string): void {
+	if(blockCache) {
+		return;
+	}
 	localStorage.removeItem(groupName);
 }
 
-export function clearCache():void{
+export function clearCache(): void {
+	if(blockCache) {
+		return;
+	}
 	localStorage.clear();
+}
+
+export function setBlockCache(block: boolean): void {
+	blockCache = block;
+}
+
+export function isCacheBlocked(): boolean {
+	return blockCache;
 }
 
 export const exerciseSolutions = 'exercise_solutions';
